Succinct.Dictionary.Class

Documentation

class Dictionary t where

Associated Types

type Elem t :: *

Methods

size :: t -> Int

size t returns the total number of elements in t.

rank :: Elem t -> t -> Int -> Int

rank a t i returns the # of occurences of a in t up to position i for 0 < i <= size t

select :: Elem t -> t -> Int -> Int

select a t i returns the position of the ith appearance of a in t. as long as 0 < i <= rank a t (size t)

(!) :: t -> Int -> Elem t

rank0 :: (Dictionary t, Elem t ~ Bool) => t -> Int -> Int

 rank0 t i = i - rank1 t i

rank1 :: (Dictionary t, Elem t ~ Bool) => t -> Int -> Int

 rank1 t i = i - rank0 t i

rank_ :: (Dictionary t, Elem t ~ Bool) => t -> Int -> Int

rank_ t i return the number of bits to the left of position i

When i > 1:

 rank_ t i = rank t (i - 1)

The result is 0 otherwise.

select0 :: (Dictionary t, Elem t ~ Bool) => t -> Int -> Int

Given i = select0 t j

 rank0 t i = j
 rank1 t i = i - j

select1 :: (Dictionary t, Elem t ~ Bool) => t -> Int -> Int

Given i = select1 t j

 rank0 t i = i - j
 rank1 t i = j