Succinct.Dictionary.Internal.Broadword

Documentation

reverseBits :: Word64 -> Word64

reverse the bits in a word

reverseBytes :: Word64 -> Word64

byteSum :: Word64 -> Word64

Calculate the total of the unsigned bytes of a 64-bit word

byteCounts :: Word64 -> Word64

Convert a word of various bits into a word where each byte contains the count of bits in the corresponding original byte

popCount = byteSum . byteCounts

lsb :: Word64 -> Int

return the position of the least significant set bit in a word, assumes that there is a set bit

msb :: Word64 -> Int

return the position of the most significant set bit in a word, assumes that there is a set bit

leq8 :: Word64 -> Word64 -> Word64

uleq8 :: Word64 -> Word64 -> Word64

unsigned compare byte by byte, returning whether or not the result is less than or equal to the corresponding byte in the other word as the least significant bit of each byte

nonzero8 :: Word64 -> Word64

selectWord64 :: Word64 -> Int -> Int