Succinct.Dictionary.RangeMin

Documentation

data RangeMin

This provides a variant of the basic aggregation used in Grossi and Ottaviano's Range-Min tree, which is in turn a simplification of a Range Min-Max tree

Constructors

RangeMin !Int !(Vector Word64) (Vector Level)

rangeMin :: Bitwise t => t -> RangeMin