Succinct.Internal.Bit

Description

Documentation

newtype Bit

Constructors

Bit Bool

class Decode64 t where

Minimal complete definition

Nothing

Methods

decode64 :: Int -> Int -> t -> Word64

wds :: Int -> Int

wd :: Int -> Int

bt :: Int -> Int

unsafeBit :: Int -> Word64

data family Vector a

data family MVector s a

foldlMPadded :: (Vector v Bit, Vector (Packed v) Word64, PackedBits v, Monad m) => (b -> Word64 -> m b) -> b -> v Bit -> m b

Fold a Bit vector as a sequence of Word64s, with the last word padded with 0 bits if incomplete.

class PackedBits v where

Expose the underlying packed representation of a bit vector

Associated Types

type Packed v :: * -> *

Methods

packedBits :: v Bit -> Packed v Word64