Succinct.Internal.Delta

Description

Documentation

data Delta

Δ-Range-Min

This provides a semigroup-based variant of the basic aggregation used in Grossi and Ottaviano's Range-Min tree, which is in turn a simplification of a Range Min-Max tree

Constructors

Delta

Fields

excess :: !Int

delta :: !Int

nminima :: !Int

minima :: Delta -> Int

bool :: Bool -> Delta

bits :: FiniteBits a => a -> Delta

byte :: Word8 -> Delta

Look up the Delta for a Word8 via LUTs

e8 :: Word8 -> Int

d8 :: Word8 -> Int

n8 :: Word8 -> Int