Succinct.Sequence

Documentation

>>> :{
let abracadabraFind 'a' = const False
    abracadabraFind 'b' = odd
    abracadabraFind 'c' = (> 0)
    abracadabraFind 'd' = even
    abracadabraFind 'r' = const True
:}

>>> :{
let bin = LabelledBin ()
    tip = LabelledTip
    abracadabraEncoding =
      bin (bin (tip 'a') (bin (tip 'b') (tip 'c'))) (bin (tip 'd') (tip 'r'))
:}

>>> :{
let (t, f) = (True, False)
    tree = LabelledBin (build [f, f, t, f, f, f, t, f, f, t, f])
           -- a & (b & c)
           (LabelledBin (build [f, t, f, t, f, f, t, f]) (LabelledTip 'a')
            -- b & c
            (LabelledBin (build [f, t, f]) (LabelledTip 'b') (LabelledTip 'c')))
           -- d & r
           (LabelledBin (build [t, f, t]) (LabelledTip 'd') (LabelledTip 'r'))
    exampleWaveletTree :: WaveletTree Rank9 Char
    exampleWaveletTree = WaveletTree tree abracadabraFind
:}

Inefficient foldable

>>> F.toList exampleWaveletTree
"abracadabra"

>>> [(c, map (select c exampleWaveletTree) [1.. rank c exampleWaveletTree (size exampleWaveletTree)]) | c <- "abcdr"]
[('a',[1,4,6,8,11]),('b',[2,9]),('c',[5]),('d',[7]),('r',[3,10])]

>>> [(c, map (rank c exampleWaveletTree) [0.. size exampleWaveletTree]) | c <- "abcdr"]
[('a',[0,1,1,1,2,2,3,3,4,4,4,5]),('b',[0,0,1,1,1,1,1,1,1,2,2,2]),('c',[0,0,0,0,0,1,1,1,1,1,1,1]),('d',[0,0,0,0,0,0,0,1,1,1,1,1]),('r',[0,0,0,1,1,1,1,1,1,1,2,2])]

data WaveletTree f a

Constructors

WaveletTree

Fields

bits :: Labelled f a

alphabet :: a -> Int -> Direction

For a given level, is the element to the right or to the left?

newtype Encoding a

Constructors

Encoding

Fields

runEncoding :: Labelled () a

buildOptimizedAlphabeticalSearchTree :: forall a n. (Show a, Eq a, Show n, Ord n, Num n, Bounded n) => [(a, n)] -> Labelled () a

huTucker :: (Bounded n, Eq a, Num n, Ord n, Show a, Show n) => [((Int, a), n)] -> Labelled () a

validHuTuckerTree :: (Eq a, Ord a) => Labelled () a -> Bool