Succinct.Tree.Binary

Basic binary tree

data Binary

Constructors

Bin Binary Binary

Tip

Conversion

jacobson :: Binary -> [Bool]

Jacobson's encoding of a binary tree

fromBinary :: Binary -> Rank9

toBinary :: Ranked t => t -> Binary

 toBinary (fromBinary t) = t

binary :: Ranked t => Zipper t -> Binary

Succinct binary tree zippers

data Zipper t

binary tree-zipper a la Jacobson

Constructors

Zipper !Int t

Operations

root :: t -> Zipper t

The root of our tree is at position 1

top :: Zipper t -> Bool

Is this node the root?

parent :: Select1 t => Zipper t -> Zipper t

parent t i returns the parent for any node i in a Binary tree, other than the root. e.g. where top i returns False

children :: Ranked t => Zipper t -> Maybe (Zipper t, Zipper t)

left :: Ranked t => Zipper t -> Zipper t

left t i returns the left child of a node i given bin t i

This is relatively unsafe. children provides a safer access pattern

right :: Ranked t => Zipper t -> Zipper t

right t i returns the right child of a node i given bin t i

 right m i = left m i + 1

This is relatively unsafe. children providers a safer access pattern

tip :: Access Bool t => Zipper t -> Bool

Is this node a Tip?

bin :: Access Bool t => Zipper t -> Bool

Is this node a Bin?