Succinct.Tree.LOUDS

Description

Delpratt, Rahman and Raman's double numbering scheme for LOUDS

Basic rose tree

newtype Rose

Constructors

Rose [Rose]

Conversion

rose :: Select0 t => Zipper t -> Rose

Extract a given sub-tree

toRose :: Select0 t => t -> Rose

 toRose . fromRose = id

fromRose :: Rose -> Rank9

Convert a finite Rose to Rank9

louds :: Rose -> [Bool]

Jacobson's 1-based LOUDS

Visit every node in level order. Represent each node with n children by (1^n)0.

We add an extra 10 "superroot" to the top of the tree to avoid corner cases.

LOUDS zipper

data Zipper t

Zipper i j t stores i in 1..2n, the position of a 1 in the LOUDS structure t along with j = rank0 t i.

All combinators in this module preserve these invariants.

Constructors

Zipper !Int !Int t

Operations

root :: t -> Zipper t

The root of our succinct tree.

index :: Zipper t -> Int

The index of this node in level order, starting at 1 for the root.

top :: Zipper t -> Bool

Is this node the root?

parent :: Select1 t => Zipper t -> Zipper t

The parent of any node i /= root, obtained by a legal sequence of operations.

children :: Select0 t => Zipper t -> [Zipper t]

positions of all of the children of a node

next :: Access Bool t => Zipper t -> Maybe (Zipper t)

next sibling, if any