Succinct.Dictionary.Class

Documentation

class Dictionary t where

Associated Types

type Elem t :: *

Methods

size :: t -> Int

size t returns the total number of elements in t.

rank :: Elem t -> t -> Int -> Int

rank a t i returns the # of occurences of a in t up to position i for 0 < i <= size t

select :: Elem t -> t -> Int -> Int

select a t i returns the position of the ith appearance of a in t. as long as 0 < i <= rank a t (size t)

(!) :: t -> Int -> Elem t