Succinct.Dictionary.Class

Documentation

class Access a t | t -> a where

Methods

size :: t -> Int

size t returns the total number of elements in t.

(!) :: t -> Int -> a

class Bitwise t where

Methods

bitwise :: t -> Vector Bit

newtype Bit

Constructors

Bit Bool

class Access a t => Dictionary a t | t -> a where

Methods

rank :: a -> t -> Int -> Int

rank a t i returns the # of occurences of a in t up to position i for 0 < i <= size t

select :: a -> t -> Int -> Int

select a t i returns the position of the ith appearance of a in t. as long as 0 < i <= rank a t (size t)

class Select0 t where

Many structures that do not support arbitrary rank can support a limited notion of select.

Methods

select0 :: t -> Int -> Int

class Select1 t where

Many structures that do not support arbitrary rank can support a limited notion of select.

Methods

select1 :: t -> Int -> Int

class (Select0 t, Select1 t, Dictionary Bool t) => Ranked t where

a classic bit-vector-based succinct indexed dictionary

For 1 <= i <= size t:

 rank0 t i + rank1 t i = i

Given i = select1 t j

 rank0 t i = i - j
 rank1 t i = j

Given i = select0 t j

 rank0 t i = j
 rank1 t i = i - j

Methods

rank0 :: Ranked t => t -> Int -> Int

 rank0 t i = i - rank1 t i
 rank0 = rank False

rank1 :: Ranked t => t -> Int -> Int

 rank1 t i = i - rank0 t i
 rank1 = rank True

rank_ :: Ranked t => t -> Int -> Int

rank_ t i return the number of bits to the left of position i

When i > 1:

 rank_ t i = rank1 t (i - 1)

The result is 0 otherwise.