Succinct.Dictionary.RangeMin

Description

Documentation

data RangeMin

This provides a variant of the basic aggregation used in Grossi and Ottaviano's Range-Min tree, which is in turn a simplification of a Range Min-Max tree

Constructors

RangeMin

Fields

_rangeMinSize :: !Int

_rangeMinRaw :: !(Vector Word64)

_rangeMinLevels :: Vector Level

rangeMin :: Bitwise t Vector => t -> RangeMin

forwardSearch :: RangeMin -> Int -> Int -> Int

forwardSearch r i d finds the minimum j > i such that excess r i j = d

backwardSearch :: RangeMin -> Int -> Int -> Int

forwardSearch r i d finds the maximum j < i such that excess r i j = d

findClose :: RangeMin -> Int -> Int

findOpen :: RangeMin -> Int -> Int

enclose :: RangeMin -> Int -> Int

levelAncestor :: RangeMin -> Int -> Int -> Int