Succinct.Tree.Binary

Succinct binary trees

data Tree

A binary tree

Constructors

Bin Tree Tree

Tip

newtype Binary

Jacobson's encoding of a succinct binary tree

Constructors

Binary Rank9

fromTree :: Tree -> Binary

toTree :: Binary -> Tree

 toTree (fromTree t) = t

Positions

root :: Int

The root of our tree is at position 1

parent :: Binary -> Int -> Int

parent t i returns the parent for any node i in a Binary tree, such that i /= root

left :: Binary -> Int -> Int

left t i returns the left child of a node i given bin t i

right :: Binary -> Int -> Int

right t i returns the right child of a node i given bin t i

 right m i = left m i + 1

tip :: Binary -> Int -> Bool

Is this node a Tip?

bin :: Binary -> Int -> Bool

Is this node a Bin?

tree :: Binary -> Int -> Tree